+++ project/__init__.py
@@ -1,34 +1,26 @@
 #!/usr/bin/env python3
-from flask import Flask
-from flask_login import LoginManager
 from flask_sqlalchemy import SQLAlchemy
-from sqlalchemy import Table, Column, Integer, ForeignKey
-from sqlalchemy.orm import relationship
-from sqlalchemy.ext.declarative import declarative_base
+from flask_login import LoginManager
+from flask import Flask
 
-app = Flask(__name__)
-db = SQLAlchemy(app)
+db = SQLAlchemy()
+login_manager = LoginManager()
+login_manager.login_view = 'main.login'
 
 
 def create_app():
+    app = Flask(__name__)
     app.config['DEBUG'] = True
     app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///database.db'
     app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-    app.config['SECRET_KEY'] = 'secret'
+    app.config['SECRET_KEY'] = 'secret_key'
 
     db.init_app(app)
-
-    from .main import main as main_blueprint
-    app.register_blueprint(main_blueprint)
-
-    login_manager = LoginManager()
-    login_manager.login_view = 'main.login'
     login_manager.init_app(app)
 
-    from .models import User
-
-    @login_manager.user_loader
-    def load_user(username):
-        return User.query.get(int(username))
+    from .main import main as main_blueprint
+    from .auth import auth as auth_blueprint
+    app.register_blueprint(main_blueprint)
+    app.register_blueprint(auth_blueprint)
 
     return app
