+++ project/templates/profile.html
@@ -2,7 +2,6 @@
 {% block content %}
 <head>
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/bulma/0.7.2/css/bulma.css" />
-
     <link rel="stylesheet" href="{{ url_for('static', filename='style.css') }}">
 </head>
     <section class="hero is-primary">
